refactor: restore the three-section layout on the members page
The page is back to Invite people, Pending invitations, Members — three
aside sections in that order, as it was before this branch. Inviting is an
inline form again rather than a modal, so the primary action is visible
without a click and matches the other workspace settings sections.

Keeps what the branch added: batch invites parsed from one textarea with a
single role, the volume rate limit, the invite link, and expiry on pending
rows. Only the chrome moves.

Section copy is trimmed to one line each; the "Please provide the email
addresses..." notice restated the field above it and is gone.

The invite path is a plain invitePeople() method, because an action nested
in a form schema is not addressable by name from outside it — that is what
main did too, and the tests call it the same way.
