+++ app/Filament/Pages/Team/Members.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Filament\Pages\Team;
 
 use App\Filament\Pages\Concerns\HasWorkspaceSettingsNavigation;
+use App\Livewire\App\Teams\InviteTeamMembers;
 use App\Livewire\App\Teams\PendingTeamInvitations;
 use App\Livewire\App\Teams\TeamMembers;
 use App\Models\Team;
@@ -64,6 +65,8 @@ final class Members extends Page
         $tenant = Filament::getTenant();
 
         return $schema->components([
+            Livewire::make(InviteTeamMembers::class)
+                ->data(['team' => $tenant]),
             Livewire::make(PendingTeamInvitations::class)
                 ->data(['team' => $tenant]),
             Livewire::make(TeamMembers::class)
