+++ app/Livewire/App/Teams/TeamMembers.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace App\Livewire\App\Teams;
 
-use App\Actions\Jetstream\InviteTeamMember;
 use App\Actions\Jetstream\RemoveTeamMember;
-use App\Actions\Jetstream\UpdateInviteLinkSettings;
 use App\Actions\Jetstream\UpdateTeamMemberRole;
 use App\Enums\TeamRole;
 use App\Livewire\BaseLivewireComponent;
@@ -16,9 +14,6 @@ use Closure;
 use Filament\Actions\Action;
 use Filament\Facades\Filament;
 use Filament\Forms\Components\Radio;
-use Filament\Forms\Components\Select;
-use Filament\Forms\Components\Textarea;
-use Filament\Infolists\Components\TextEntry;
 use Filament\Tables;
 use Filament\Tables\Table;
 use Illuminate\Auth\Access\AuthorizationException;
@@ -27,37 +22,19 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Gate;
-use Illuminate\Support\Facades\RateLimiter;
 use Illuminate\Validation\ValidationException;
 use Laravel\Jetstream\Jetstream;
 
 /**
- * The workspace's people directory. Pending invitations are a separate,
- * unpaginated worklist rendered by PendingTeamInvitations — merging the two
- * into one paginated table buried the actionable set behind pagination, which
- * is why Twenty and Slack keep them apart too.
+ * The workspace's people directory. Inviting lives in InviteTeamMembers and
+ * pending invitations in PendingTeamInvitations — three sections down the page,
+ * so the roster stays a roster and the actionable sets are not buried inside
+ * it. Merging them into one paginated table hid pending invites behind
+ * pagination, which is why Twenty and Slack keep them apart too.
  */
 final class TeamMembers extends BaseLivewireComponent implements Tables\Contracts\HasTable
 {
     use Tables\Concerns\InteractsWithTable;
-
-    /**
-     * Bounds one invite submission — generous for real bulk-inviting a team,
-     * but stops a single authorized call from queuing an unbounded number of
-     * Mail sends (the onboarding wizard's invite step caps at 5 for the same
-     * reason; this screen is used long after onboarding, so it allows more).
-     */
-    private const int MAX_INVITES_PER_SUBMISSION = 10;
-
-    /**
-     * Bounds cumulative invite volume per actor within the window below,
-     * independent of how many separate submissions it takes to reach it —
-     * `rateLimit()` alone only throttles the number of calls, not the number
-     * of emails each call can queue.
-     */
-    private const int MAX_INVITES_PER_WINDOW = 20;
-
-    private const int INVITE_WINDOW_SECONDS = 60;
 
     public Team $team;
 
@@ -72,10 +49,6 @@ final class TeamMembers extends BaseLivewireComponent implements Tables\Contract
             ->query($this->membersQuery(...))
             ->paginated(false)
             ->defaultSort('name')
-            ->headerActions([
-                $this->invitePeopleAction(),
-                $this->manageInviteLinkAction(),
-            ])
             // Split rather than discrete columns: this is a short roster, and a
             // header row over two fields reads as table chrome around a list.
             ->columns([
@@ -176,166 +149,6 @@ final class TeamMembers extends BaseLivewireComponent implements Tables\Contract
         }
 
         return $registeredRole->name;
-    }
-
-    private function invitePeopleAction(): Action
-    {
-        return Action::make('invitePeople')
-            ->label(__('teams.actions.invite_people'))
-            ->icon('heroicon-m-user-plus')
-            ->visible(fn (): bool => Gate::check('addTeamMember', $this->team))
-            ->modalHeading(__('teams.modals.invite_people.heading'))
-            ->modalWidth('lg')
-            ->modalSubmitActionLabel(__('teams.actions.send_invitations'))
-            ->schema([
-                Textarea::make('emails')
-                    ->label(__('teams.form.emails.label'))
-                    ->placeholder(__('teams.form.emails.placeholder'))
-                    ->helperText(__('teams.form.emails.helper'))
-                    ->rows(4)
-                    ->required()
-                    ->rule(fn (): Closure => function (string $attribute, mixed $value, Closure $fail): void {
-                        $emails = $this->parseEmails(is_string($value) ? $value : '');
-
-                        if ($emails === []) {
-                            $fail(__('teams.validation.no_valid_emails'));
-
-                            return;
-                        }
-
-                        if (count($emails) > self::MAX_INVITES_PER_SUBMISSION) {
-                            $fail(__('teams.validation.too_many_invites', ['max' => self::MAX_INVITES_PER_SUBMISSION]));
-                        }
-                    }),
-                Select::make('role')
-                    ->label(__('teams.form.invite_as.label'))
-                    ->options(fn (): array => $this->assignableRoles())
-                    ->in(fn (): array => array_keys($this->assignableRoles()))
-                    ->default(TeamRole::Editor->value)
-                    ->selectablePlaceholder(false)
-                    ->required(),
-            ])
-            ->action(function (array $data): void {
-                $this->sendInvitations(
-                    $this->parseEmails((string) $data['emails']),
-                    (string) $data['role'],
-                );
-            });
-    }
-
-    /**
-     * Split a pasted block into addresses. Attio accepts one textarea and lets
-     * people paste from a spreadsheet or a mail client, so anything that can
-     * separate addresses in those sources counts: commas, semicolons, and any
-     * whitespace including newlines.
-     *
-     * @return list<string>
-     */
-    private function parseEmails(string $input): array
-    {
-        $parts = preg_split('/[\s,;]+/', trim($input)) ?: [];
-
-        return array_values(array_unique(array_filter(
-            array_map(trim(...), $parts),
-            fn (string $email): bool => $email !== '',
-        )));
-    }
-
-    /**
-     * @param  list<string>  $emails
-     */
-    private function sendInvitations(array $emails, string $role): void
-    {
-        if ($emails === []) {
-            return;
-        }
-
-        // Volume-based, not call-based: `rateLimit()` would only throttle how
-        // often this method runs, not how many emails a single run queues —
-        // `maxItems()` on the repeater already bounds one submission, this
-        // bounds cumulative volume across submissions from the same actor.
-        $rateLimitKey = 'invite-team-members:'.$this->authUser()->id;
-
-        if (RateLimiter::tooManyAttempts($rateLimitKey, self::MAX_INVITES_PER_WINDOW)) {
-            $this->sendNotification(
-                __('teams.notifications.invite_rate_limited.title'),
-                __('teams.notifications.invite_rate_limited.body', [
-                    'seconds' => RateLimiter::availableIn($rateLimitKey),
-                ]),
-                'danger',
-            );
-
-            return;
-        }
-
-        RateLimiter::increment($rateLimitKey, self::INVITE_WINDOW_SECONDS, count($emails));
-
-        $failures = [];
-        $sent = 0;
-
-        foreach ($emails as $email) {
-            try {
-                resolve(InviteTeamMember::class)->invite($this->authUser(), $this->team, $email, $role);
-
-                $sent++;
-            } catch (ValidationException $exception) {
-                $failures[] = "{$email}: {$exception->validator->errors()->first()}";
-            }
-        }
-
-        if ($sent > 0) {
-            $this->sendNotification(__('teams.notifications.team_invitation_sent.success'));
-            $this->dispatch('teamInvitationSent');
-        }
-
-        if ($failures !== []) {
-            $this->sendNotification(
-                __('teams.notifications.some_invites_failed.title'),
-                implode("\n", $failures),
-                'warning',
-            );
-        }
-    }
-
-    private function manageInviteLinkAction(): Action
-    {
-        return Action::make('manageInviteLink')
-            ->label(__('teams.actions.invite_link'))
-            ->icon('heroicon-m-link')
-            ->color('gray')
-            ->visible(fn (): bool => Gate::check('addTeamMember', $this->team))
-            ->schema([
-                TextEntry::make('url')
-                    ->label(__('teams.invite_link.url'))
-                    ->state(fn (): string => route('teams.join', ['token' => $this->team->invite_link_token]))
-                    ->copyable(),
-                Select::make('invite_link_default_role')
-                    ->label(__('teams.invite_link.default_role'))
-                    ->options(fn (): array => $this->assignableRoles())
-                    ->in(fn (): array => array_keys($this->assignableRoles()))
-                    ->default(fn (): string => $this->team->invite_link_default_role)
-                    ->required(),
-            ])
-            ->extraModalFooterActions([
-                Action::make('rotateInviteLink')
-                    ->label(__('teams.actions.rotate_invite_link'))
-                    ->color('danger')
-                    ->requiresConfirmation()
-                    ->action(function (): void {
-                        resolve(UpdateInviteLinkSettings::class)->rotate($this->authUser(), $this->team);
-
-                        $this->sendNotification(__('teams.notifications.invite_link_rotated.success'));
-                    }),
-            ])
-            ->action(function (array $data): void {
-                resolve(UpdateInviteLinkSettings::class)->update(
-                    $this->authUser(),
-                    $this->team,
-                    (string) $data['invite_link_default_role'],
-                );
-
-                $this->sendNotification();
-            });
     }
 
     private function updateTeamRoleAction(): Action
