+++ resources/views/livewire/app/teams/invite-team-members.blade.php
@@ -1,0 +1,5 @@
+<div>
+    {{ $this->form }}
+
+    <x-filament-actions::modals/>
+</div>
